Recipes Finished!
Well, at least the AE/Extra Cell ones. All the recipes where my items
made sense to include, I remade in this mod with both AE and my variants
of the items. Also in this build is the first compiled version, ready
for use!

Still needs commenting, but I'm happy with my progress and completion
enough to release a build.